Move flag types to a new flags.go file
This is to reduce the complexity of the sandboxfs.go file by isolating
high-level flag definitions into their own file.
