+++ cmd/sandboxfs/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "errors.go",
+        "flags.go",
         "profiling.go",
         "sandboxfs.go",
     ],
